Update GetElected.java
This will be the final change to this version. I have been headed down the wrong path for a while with this little learner program. I need to create multiple methods and classes and calls and use gobs and gobs of tools in order to make a simple top to bottom program easier to write ... right.

As it stands, the program works but it doesn't go through an election cycle because I can't get the damn thing to recognize the LABEL that it uses everywhere else. Why? Because my coding is too simple to be trusted to work. I don't have enough complexity. I'm thinking like BASIC. Whatever. I'll load up MAVEN and GRADEL and KOTNER and ANT and CUCUMBER and SPRING and RINGADINGDING to make this little program simpler to run ... somehow ...